requirements.txt
pyexiftool==0.5.5
inflect==5.6.0
pydantic==1.10.7
nltk==3.8.1
tqdm==4.65.0